renamed internal to internals
