+++ internals/sql/repository/ImageScanDeployInfoRepository.go
