+++ internals/sql/repository/ImageScanHistoryRepository.go
