+++ internals/sql/repository/ImageScanObjectMetaRepository.go
