+++ internals/sql/repository/ImageScanResultRepository.go
