+++ internals/sql/repository/RegistryIndexMappingRepository.go
