+++ internals/sql/repository/ScanStepConditionMappingRepository.go
