+++ internals/sql/repository/UserRepository.go
@@ -6,7 +6,7 @@ package repository
 
 import (
 	"fmt"
-	"github.com/devtron-labs/image-scanner/internal/sql/bean"
+	"github.com/devtron-labs/image-scanner/internals/sql/bean"
 	"github.com/go-pg/pg"
 	"go.uber.org/zap"
 	"time"
