+++ internals/util/ErrorUtil.go
@@ -17,12 +17,12 @@ func (e *ApiError) Error() string {
 	return e.InternalMessage
 }
 
-// default internal will be set
+// default internals will be set
 func (e *ApiError) ErrorfInternal(format string, a ...interface{}) error {
 	return &ApiError{InternalMessage: fmt.Sprintf(format, a...)}
 }
 
-//default user message will be set
+// default user message will be set
 func (e ApiError) ErrorfUser(format string, a ...interface{}) error {
 	return &ApiError{InternalMessage: fmt.Sprintf(format, a...)}
 }
