+++ pkg/security/ImageScanService.go
@@ -9,11 +9,11 @@ import (
 	"github.com/Knetic/govaluate"
 	"github.com/caarlos0/env"
 	"github.com/devtron-labs/image-scanner/common"
-	"github.com/devtron-labs/image-scanner/internal/sql/bean"
-	"github.com/devtron-labs/image-scanner/internal/sql/repository"
-	cliUtil "github.com/devtron-labs/image-scanner/internal/step-lib/util/cli-util"
-	commonUtil "github.com/devtron-labs/image-scanner/internal/step-lib/util/common-util"
-	httpUtil "github.com/devtron-labs/image-scanner/internal/step-lib/util/http-util"
+	"github.com/devtron-labs/image-scanner/internals/sql/bean"
+	"github.com/devtron-labs/image-scanner/internals/sql/repository"
+	cliUtil "github.com/devtron-labs/image-scanner/internals/step-lib/util/cli-util"
+	commonUtil "github.com/devtron-labs/image-scanner/internals/step-lib/util/common-util"
+	httpUtil "github.com/devtron-labs/image-scanner/internals/step-lib/util/http-util"
 	"github.com/go-pg/pg"
 	"github.com/optiopay/klar/clair"
 	"github.com/quay/claircore"
