+++ wire_gen.go
@@ -10,9 +10,9 @@ import (
 	"github.com/devtron-labs/common-lib/monitoring"
 	"github.com/devtron-labs/common-lib/pubsub-lib"
 	"github.com/devtron-labs/image-scanner/api"
-	"github.com/devtron-labs/image-scanner/internal/logger"
-	"github.com/devtron-labs/image-scanner/internal/sql"
-	"github.com/devtron-labs/image-scanner/internal/sql/repository"
+	"github.com/devtron-labs/image-scanner/internals/logger"
+	"github.com/devtron-labs/image-scanner/internals/sql"
+	"github.com/devtron-labs/image-scanner/internals/sql/repository"
 	"github.com/devtron-labs/image-scanner/pkg/clairService"
 	"github.com/devtron-labs/image-scanner/pkg/grafeasService"
 	"github.com/devtron-labs/image-scanner/pkg/klarService"
